requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
fastapi-users[sqlalchemy]==14.0.1
sqlalchemy[asyncio]>=2.0
asyncpg==0.30.0
python-multipart==0.0.20
python-dotenv==1.0.1
email-validator==2.1.1
cassandra-driver==3.29.2
kubernetes==31.0.0
websockets>=12.0
python-jose[cryptography]==3.3.0

